Introducing first version of md4 algorithm. CAUTION: it is currently untested!
I will write some tests cases included in the paper later today
or tomorrow. More tests and (probable) debugging tomorrow.

It uses (external) `bitarray` class for operations on bits.
Later, we should check if this class can work as a stream instead of
array, because for larger files it may struggle to contain it
in the RAM all at once. Built-in `bytearray` could also be
considered as an alternative.